// src/app/models/usuario.ts
export class Usuario {
    _id: number;
    name: String;
    last_name: String;
    tercero: any;
    email: String;
    telefono: String;
    imagen: String;
    Password: String;
    analista: Boolean;
    infofinanciera: Boolean;
    recepcionM: Boolean;
    firma: String;
    firmaOriginal: String;
    firmaWidth: number;
    firmaHeight: number;
    rol: string;
}